+++ mm2src/coins/rpc_command/mod.rs
@@ -8,3 +8,4 @@ pub mod init_create_account;
 pub mod init_scan_for_new_addresses;
 pub mod init_withdraw;
 #[cfg(not(target_arch = "wasm32"))] pub mod lightning;
+pub mod tendermint;
